Added util class with reusable date and time functions
